+++ workspace/sort-packages/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/sort-packages
 
+## 5.0.5
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+
 ## 5.0.4
 
 ### Patch Changes